Change divs containing numbers to buttons
To begin with the implementation of code to the calculator's interface,
instead of using divs for each number in the html, now buttons are
being used as they facilitate the interaction.

While technically both divs and buttons would work, buttons should be
used to improve accessibility.

This change also applies to opperands.