Create button logic for the numbers
By using similar code to the one in the Etch-a-sketch project, all the
buttons inside a specific div are selected and stored inside a const
value. Then each one of the buttons can have specific actions by
checking the text inside it.

This will be extended to operands as well.