Add logic to make operations
A boolean "isSecondValue" determines if we're storing the first or the
second value, this is initiated as false when the code runs as we're
always starting with the first one.

Since having one number stored is necessary to operate, the boolean is
not changed unless an operator is selected, in that case, we can
select the second number.

Once the = button is pressed the bool will be set to false and the
function to make the operation will be called, then the result is
consoled log for testing.