Store multiple numbers in each value
By using an array that stores a set of temporary values, every time a
number is pressed in the calculator, it gets added to the value until
an operator gets chosen.

This is done by using the .join() method and converting the returned
value of the method into a number then storing it inside valueA or
valueB.

To prevent the second number from being the same as the first, the
temp array is cleared after an operator is chosen. Likewise, using AC
clears it to prevent issues.