Evaluate only two numbers at a time
To achieve this, it was first necessary to initialize the values of both
numbers as empty <div> elements, meaning they contain an empty string ('').

This step is important because the code checks whether valueB already
contains a number each time an operator is clicked.

To ensure accurate results, the program must keep track of which
operator was selected before and after each one is pressed. An array
called operatorContainer stores both the current and the previous
operators. A function limits the array’s length to a maximum of two
elements by using the shift method which eliminates an item from the
beginning of the array.

If valueB is not empty and the isSecondNumber condition is true, and the
user presses an operator for the third time, the program performs an
operation using the previously stored operator. The result is then
stored in valueA and displayed as the first number, while the valueB
<div> is cleared.

By using a boolean variable (isEqual), the program can detect whether
the user intends to continue adding another number to the operation or
display the final result of the two numbers.