Modify display for better readability
The divs containing the numbers and the results have been put in
different lines so it's easier for the user to read the information on
the calculator.

This was done by adding two containers inside the display div, one for
the user numbers and operator and one for the results, in the middle
the = sign is left outside these containers.