Add new decimal button logic
A new boolean variable was added to ensure the decimal button (.) can
only be used once per number input.

When the decimal button is pressed and the condition allows it (i.e.
the boolean is true), a . is added to the current input. This means any
numbers entered after it will be treated as part of a decimal number.
After that, the boolean is set to false, preventing additional decimal
points from being added.

Once an operator is pressed (like +, -, etc.), the boolean is reset to
true, allowing a decimal point to be used again, only once, for the
next number.